Move calibration logic into separate class
